refactor(fleet): extract RPC/CRUD handlers to fleet-rpc-handlers.ts (P4.1)
Move ~355 lines of fleet-side RPC and HTTP CRUD handlers out of FleetManager:
- handleScheduleCrud / handleDecisionCrud / handleTaskCrud
- handleSetDisplayName / handleSetDescription
- handleScheduleCrudHttp / handleDecisionCrudHttp / handleTaskCrudHttp
- handleSetDisplayNameHttp / handleSetDescriptionHttp
- summarizeToolCall (pure)
- resolveDisplayName (pure)

Same pattern as outbound-handlers.ts: a narrow `RpcHandlersContext` interface
+ a `rpcHandlers` dispatch dictionary. FleetManager keeps thin forwarders for
the 5 *Http methods (required by AgentEndpointContext) and resolveDisplayName;
internal IPC dispatch sites call rpcHandlers.X(this, ...) directly.

Behaviour and IPC wire format unchanged.

P4.1 step 3 of 4.
